techIt/requirements.txt
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10
tqdm==4.65.0